sdks/java/io/elasticsearch7/src/test/java/org/apache/beam/sdk/io/elasticsearch/SerializableValueInSingleWindowTest.java
/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.apache.beam.sdk.io.elasticsearch;

import static org.apache.beam.sdk.io.elasticsearch.utils.Elasticsearch7IOTestUtils.buildIndexRequest;
import static org.assertj.core.api.Assertions.assertThat;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.io.ObjectInputStream;
import java.io.ObjectOutputStream;
import org.apache.beam.sdk.transforms.windowing.GlobalWindow;
import org.apache.beam.sdk.transforms.windowing.PaneInfo;
import org.elasticsearch.action.index.IndexRequest;
import org.joda.time.Instant;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.junit.runners.JUnit4;

@RunWith(JUnit4.class)
public class SerializableValueInSingleWindowTest {
  @Test
  @SuppressWarnings("unchecked")
  public void defaultTest() throws IOException, ClassNotFoundException {
    /** Serialize the value. */
    SerializableValueInSingleWindow<IndexRequest> value =
        SerializableValueInSingleWindow.of(
            buildIndexRequest(), Instant.now(), GlobalWindow.INSTANCE, PaneInfo.NO_FIRING);

    ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
    ObjectOutputStream objectOutputStream = new ObjectOutputStream(byteArrayOutputStream);

    objectOutputStream.writeObject(value);

    /** Deserialize the result */
    ByteArrayInputStream byteArrayInputStream =
        new ByteArrayInputStream(byteArrayOutputStream.toByteArray());
    ObjectInputStream objectInputStream = new ObjectInputStream(byteArrayInputStream);

    SerializableValueInSingleWindow<IndexRequest> deserializedValue =
        (SerializableValueInSingleWindow<IndexRequest>) objectInputStream.readObject();

    assertThat(deserializedValue).isEqualTo(value);
  }
}
